# golaco-api/src/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Golaço API
  description: 'Golaço API, endpoints criados pela dupla [Luis Carlos Zancanela](https://git.gft.com/lsza) e [Vinicius Santos Silva](https://git.gft.com/vsso) para serem utilizados pelo Golaço APP'
  termsOfService: http://swagger.io/terms/
  contact:
    name: Luis Zancanela
    email: luis.zancanela@gft.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: http://localhost:3000/v1
...
paths:
...
  /transactions/deposit:
    post:
      tags:
      - Transactions
      - HTTP Method POST
      - Only Own User
      summary: 'Transação de depósito'
      description: 'Realizar transação de depósito de recurso'
      operationId: depositTransaction
      requestBody:
        description: 'Pessoa e valor para a operação'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTransaction'
      responses:
        201: 
          description: 'Transação realizada com sucesso'
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ResponseTransaction'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        409:
           $ref: '#/components/responses/Conflict'
      security:
        - bearerAuth: []
  /transactions/withdraw:
    post:
      tags:
      - Transactions
      - HTTP Method POST
      - Only Own User
      summary: 'Transação de retirada'
      description: 'Realizar transação de retirada de recurso'
      operationId: withdrawTransaction
      requestBody:
        description: 'Pessoa e valor para a operação'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTransaction'
      responses:
        201: 
          description: 'Transação realizada com sucesso'
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ResponseTransaction'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        409:
           $ref: '#/components/responses/Conflict'
      security:
        - bearerAuth: []
  /transactions/transfer:
    post:
      tags:
      - Transactions
      - HTTP Method POST
      - Only Own User
      summary: 'Transação de retirada'
      description: 'Realizar transação de retirada de recurso'
      operationId: transferTransaction
      requestBody:
        description: 'Pessoa e valor para a operação'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTransfer'
      responses:
        201: 
          description: 'Transação realizada com sucesso'
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ResponseTransaction'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        409:
           $ref: '#/components/responses/Conflict'
      security:
        - bearerAuth: []
  /teams:
    get:
      tags:
      - Teams
      - HTTP Method GET
      - Role ADMIN
      - Role USER
      summary: 'Pesquisa todos os times cadastrados no sistema'
      description: 'Pesquisar todos os times e retorna com paginação'
      operationId: findAllTeams
      parameters:
        - in: query
          name: take
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
          required: false
          description: 'Quantidade de itens por página.'
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
          description: 'Número da página.'
      responses:
        200:
          description: 'Operação realizada com sucesso.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseTeam'
        401:
          description: 'Unauthorized - Não está autenticado'
          content: {}
        403:
          description: 'Forbidden - Acesso não autorizado'
          content: {}
      security:
        - bearerAuth: []
    post:
      tags:
      - Teams
      - HTTP Method POST
      - Role ADMIN
      summary: 'Adiciona novo time'
      operationId: addTeam
      requestBody:
        description: 'Dados de conta para adicionar ao sistema'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTeam'
        required: true
      responses:
        201:
          description: 'Created - Novo time salvo com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseTeam'
        400:
          description: 'BadRequest - Conteúdo inválido'
          content: {}
        401:
          description: 'Unauthorized - Não está autenticado'
          content: {}
        403:
          description: 'Forbidden - Acesso não autorizado'
          content: {}
        409:
          description: 'Conflict - Algum dado enviado está inválido'
          content: {}
      security:
        - bearerAuth: []
  /teams/{id}:
    get:
      tags:
      - Teams
      - HTTP Method GET
      - Role ADMIN
      - Role USER
      summary: 'Pesquisa time por id'
      operationId: getTeamById
      parameters:
      - name: id
        in: path
        description: 'Id do time'
        required: true
        example: 1
        schema:
          type: integer
          format: int64
      responses:
        200: 
          description: 'Time encontrado e retornado com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseTeam'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
           $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []    
    put:
      tags:
      - Teams
      - HTTP Method PUT
      - Role ADMIN
      summary: 'Atualiza time por id'
      operationId: updateTeamById
      parameters:
      - name: id
        in: path
        description: 'Id do time'
        required: true
        example: 1
        schema:
          type: integer
          format: int64
      requestBody:
        description: 'Atributos do time para ser atualizado no sistema'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTeam'
        required: true
      responses:
        200:
          description: 'Time atualizado com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseTeam'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
           $ref: '#/components/responses/Conflict'
      security:
        - bearerAuth: []
    delete:
      tags:
      - Teams
      - HTTP Method DELETE
      - Role ADMIN
      summary: 'Excluir conta por id'
      operationId: deleteTeamById
      parameters:
      - name: id
        in: path
        description: 'Id do time'
        required: true
        example: 1
        schema:
          type: integer
          format: int64
      responses:
        204: 
          description: 'Exclusão realizada com sucesso'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
           $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []
...
  /users:
    get:
      tags:
      - Users
      - HTTP Method GET
      - Role ADMIN
      summary: 'Pesquisa todos os usuários'
      description: 'Pesquisar todos os usuários, retorno com paginação'
      operationId: findAllUsers
      parameters:
        - in: query
          name: take
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
          required: false
          description: 'Quantidade de itens por página.'
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
          description: 'Número da página.'
      responses:
        200:
          description: 'Operação realizada com sucesso.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseUser'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    post:
      tags:
      - Users
      - HTTP Method POST
      - Role ADMIN
      summary: Adiciona novo usuário
      operationId: addUser
      requestBody:
        description: 'Objeto com dados de usuário para adicionar ao sistema'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUser'
        required: true
      responses:
        201:
          description: 'Created - Novo usuário salvo com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUser'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        409:
           $ref: '#/components/responses/Conflict'
      security:
        - bearerAuth: []
  /users/{id}:
    get:
      tags:
      - Users
      - HTTP Method GET
      - Role ADMIN
      - Only Own User
      summary: 'Pequisa usuário por id'
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: 'Id do usuário'
        required: true
        example: 1
        schema:
          type: integer
          format: int64
      responses:
        200: 
          description: 'Usuário encontrado e retornado com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUser'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
           $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []    
    put:
      tags:
      - Users
      - HTTP Method PUT
      - Role ADMIN
      summary: 'Atualiza usuário por id'
      operationId: updateUserById
      parameters:
      - name: id
        in: path
        description: 'Id do usuário'
        required: true
        example: 1
        schema:
          type: integer
          format: int64
      requestBody:
        description: 'Atributos do objeto usuário para ser atualizado no sistema'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUser'
        required: true
      responses:
        200:
          description: 'Usuário atualizado com sucesso'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUser'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        409:
           $ref: '#/components/responses/Conflict'
      security:
        - bearerAuth: []
    delete:
      tags:
      - Users
      - HTTP Method DELETE
      - Role ADMIN
      summary: 'Excluir usuário por id'
      operationId: deleteUserById
      parameters:
      - name: id
        in: path
        description: 'Id do usuário'
        required: true
        example: 1
        schema:
          type: integer
          format: int64
      responses:
        204: 
          description: 'Exclusão realizada com sucesso'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
           $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []
  /users/{id}/password:
    patch:
      tags:
      - Users
      - HTTP Method PATCH
      - Only Own User
      summary: 'Atualiza senha do usuário'
      operationId: updateUserPasswordById
      parameters:
      - name: id
        in: path
        description: Id do usuário
        required: true
        example: 1
        schema:
          type: integer
          format: int64
      requestBody:
        description: 'Senha antiga e nova para atualizar o atributo do usuário no sistema'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUserPassword'
        required: true
      responses:
        204:
          description: 'Senha do usuário alterada com sucesso'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        409:
           $ref: '#/components/responses/Conflict'
      security:
        - bearerAuth: []
components:
  responses:
    BadRequest:
      description: 'BadRequest - Conteúdo inválido'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 'Unauthorized - Não está autenticado'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 'Forbidden - Acesso não autorizado'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 'O recurso especificado não foi encontrado'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 'Conflict - Algum dado enviado está inválido'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: string
        timestamp:
          type: string
        message:
          type: string
    
    #Objects
...
    RequestUser:
      type: object
      properties:
        username:
          $ref: '#/components/schemas/Username'
        password:
          $ref: '#/components/schemas/Password'
        role:
          $ref: '#/components/schemas/Role'
        personId:
          type: integer
          format: int64
          description: 'Id da pessoa vinculada'
          example: 1
      required:
        - username
        - password
    RequestUserPassword:
      type: object
      properties:
        oldPassword:
          $ref: '#/components/schemas/Password'
        newPassword:
          $ref: '#/components/schemas/Password'
      required:
        - password
    ResponseUser:
      type: object
      properties:
        id:
          type: integer
          description: Id do usuário
          format: int64
          example: 1
        username:
          $ref: '#/components/schemas/Username'
        roles:
          $ref: '#/components/schemas/Role'
        personId:
          $ref: '#/components/schemas/PersonId'
    PageResponseUser:
      type: object
      properties:
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
        data:
          type: array
          items: 
            $ref: '#/components/schemas/ResponseUser'
    RequestTransaction:
      type: object
      properties:
        amount:
          $ref: '#/components/schemas/Amount'
        personId:
          type: integer
          format: int64
          description: 'Id da pessoa que está fazendo o depósito ou retirada'
          example: 1
    RequestTransfer:
      type: object
      properties:
        name:
          type: string
          description: 'Nome da transação'
          example: 'Aposta Realizada'
        amount:
          $ref: '#/components/schemas/Amount'
        payer:
          type: integer
          format: int64
          description: 'Id da pessoa que está pagando'
          example: 1
        receiver:
          type: integer
          format: int64
          description: 'Id da pessoa que está recebendo'
          example: 2
    ResponseTransaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 'Id da transação'
          example: 1
        name:
          type: string
          description: 'Nome da transação'
          example: 'Aposta Realizada'
        amount:
          $ref: '#/components/schemas/Amount'
        dateTimeAt:
          type: string
          description: 'Data e hora da transação'
          example: '2022-09-06T15:06:29.000Z'  
        payer:
          type: integer
          format: int64
          description: 'Id da conta que paga'
          example: 1
        receiver:
          type: integer
          format: int64
          description: 'Id da conta que recebe'
          example: 2
        betId:
          type: integer
          format: int64
          nullable: true
          description: 'Id da aposta vinculada à transação'
          example: 1
        status:
          type: string
          description: 'Status da transação'
          example: 'done'
          enum: 
            ['done', 'closed']
...
    RequestTeam:
      type: object
      properties:
        name:
          type: string
          description: 'Nome do Time'
          example: 'Brasil'
        teamImage:
          type: string
          description: 'URL da imagem do time'
          example: 'http://dominio.com/folder/image.jpg'
      required:
        - name
    ResponseTeam:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 'Id do time'
          example: '1'
        name:
          type: string
          description: 'Nome do Time'
          example: 'Brasil'
        teamImage:
          type: string
          description: 'URL da imagem do time'
          example: 'http://dominio.com/folder/image.jpg'
    PageResponseTeam:
      type: object
      properties:
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
        data:
          type: array
          items: 
            $ref: '#/components/schemas/ResponseTeam'
...
    PageInfo:
      type: object
      properties:
        itemsPerPage:
          type: integer
          format: int32
          default: 100
        currentPage:
          type: integer
          format: int32
          default: 0
        currentPageAmountItems:
          type: integer
          format: int32
          default: 50
        lastPage:
          type: integer
          format: int32
          default: 0
        totalItems:
          type: integer
          format: int32
          default: 50
    
    #Fields    
    Role:
      type: string
      nullable: true
      description: 'Perfil de Usuário'
      example: 'ADMIN'
      enum: 
        ['ADMIN', 'USER']
...
    Username:
      type: string
      description: 'Nome do usuário'
      example: 'usuario'
    Password:
      type: string
      description: 'Senha do usuário'
      example: 123456
    PersonId:
      type: integer
      format: int64
      description: 'Id da pessoa vinculada'
      example: 1
...
    Amount:  
      type: string
      description: 'Quantia'
      example: '100.67'
    
      
  #Security            
  securitySchemes:
    bearerAuth:         
      type: http
      scheme: bearer
      bearerFormat: JWT